fix: handle invalid JSON files when loading scenarios from directory
